make json reports triggerable with cli argument

+++ conda-recipe/run_test.sh
@@ -34,41 +34,47 @@ fi
 echo "Start testing ..."
 return_code=0
 
+# Note: execute with argument --json-report in order to produce
+# a JSON report under folder '.pytest_reports'. Other arguments
+# will also be forwarded to pytest.
+if [[ "$*" == *"--json-report"* ]]; then
+    echo "Will produce JSON report of tests"
+    mkdir -p .pytest_reports
+    if [[ ! -z "$(ls .pytest_reports)" ]]; then
+        rm .pytest_reports/*.json
+    fi
+fi
+
 python -c "import daal4py"
 return_code=$(($return_code + $?))
 
-mkdir -p .pytest_reports
-if [[ ! -z "$(ls .pytest_reports)" ]]; then
-    rm .pytest_reports/*.json
-fi
-
 echo "Pytest run of legacy unittest ..."
 echo ${daal4py_dir}
-pytest --verbose -s ${daal4py_dir}/tests --json-report --json-report-file=.pytest_reports/legacy_report.json
+pytest --verbose -s ${daal4py_dir}/tests $@ --json-report-file=.pytest_reports/legacy_report.json
 return_code=$(($return_code + $?))
 
 echo "NO_DIST=$NO_DIST"
 if [[ ! $NO_DIST ]]; then
     echo "MPI pytest run of legacy unittest ..."
     mpirun --version
-    mpirun -n 4 pytest --verbose -s ${daal4py_dir}/tests/test*spmd*.py --json-report --json-report-file=.pytest_reports/mpi_legacy_report.json
+    mpirun -n 4 pytest --verbose -s ${daal4py_dir}/tests/test*spmd*.py $@ --json-report-file=.pytest_reports/mpi_legacy_report.json
     return_code=$(($return_code + $?))
 fi
 
 echo "Pytest of daal4py running ..."
-pytest --verbose --pyargs ${daal4py_dir}/daal4py/sklearn --json-report --json-report-file=.pytest_reports/daal4py_report.json
+pytest --verbose --pyargs ${daal4py_dir}/daal4py/sklearn $@ --json-report-file=.pytest_reports/daal4py_report.json
 return_code=$(($return_code + $?))
 
 echo "Pytest of sklearnex running ..."
-pytest --verbose --pyargs sklearnex --json-report --json-report-file=.pytest_reports/sklearnex_report.json
+pytest --verbose --pyargs sklearnex $@ --json-report-file=.pytest_reports/sklearnex_report.json
 return_code=$(($return_code + $?))
 
 echo "Pytest of onedal running ..."
-pytest --verbose --pyargs ${daal4py_dir}/onedal --json-report --json-report-file=.pytest_reports/onedal_report.json
+pytest --verbose --pyargs ${daal4py_dir}/onedal $@ --json-report-file=.pytest_reports/onedal_report.json
 return_code=$(($return_code + $?))
 
 echo "Global patching test running ..."
-pytest --verbose -s ${daal4py_dir}/.ci/scripts/test_global_patch.py --json-report --json-report-file=.pytest_reports/global_patching_report.json
+pytest --verbose -s ${daal4py_dir}/.ci/scripts/test_global_patch.py $@ --json-report-file=.pytest_reports/global_patching_report.json
 return_code=$(($return_code + $?))
 
 exit $return_code
